# Requestify ExportedSourceFile parsing