Avoid parsing the syntax tree up-front, and instead only parse it when required, which happens when either:

1. ASTGen parsing is enabled (currently disabled by default)
2. Round trip checking is enabled for a primary file (enabled by default in a debug build, except when dep scanning or doing an IDE operation)
3. We need to evaluate a macro in that file

This change therefore means that we now no longer need to parse the syntax tree for secondary files by default unless we specifically need to evaluate a macro in them (e.g if we need to lookup a member on a decl with an attached macro). And the same for primaries in release builds.

rdar://109283847