@hamishknight 2023-05-19
@swift-ci please test compiler performance