@hamishknight 2023-05-22
Seems like some flakiness is going on with the source compat suite, sometimes UPASS, sometimes XFAIL...

https://ci.swift.org/job/swift-PR-compiler-performance-macOS/63/

Comment from the run:

# Summary for main full

**Regressions found (see below)**

- [Debug-batch](#debug-batch-2af00efbbb825807)
    - [debug-batch brief](#debug-batch-brief-2af00efbbb825807)
    - [debug-batch detailed](#debug-batch-detailed-2af00efbbb825807)
- [Release](#release-2af00efbbb825807)
    - [release brief](#release-brief-2af00efbbb825807)
    - [release detailed](#release-detailed-2af00efbbb825807)



<a name="debug-batch-2af00efbbb825807">

# Debug-batch

<a name="debug-batch-brief-2af00efbbb825807">

## debug-batch brief
<details>
<summary>Regressed (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Improved (1)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Frontend.NumInstructionsExecuted | 145,026,849,632,443 | 122,931,368,255,179 | -22,095,481,377,264 | -15.24% :white_check_mark:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (1)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
LLVM.NumLLVMBytesOutput | 2,443,809,896 | 2,443,852,960 | 43,064 | 0.0%
</details>

<a name="debug-batch-detailed-2af00efbbb825807">

## debug-batch detailed
<details>
<summary>Regressed (9)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.NumSourceLinesPerSecond | 3,700,336 | 5,018,529 | 1,318,193 | 35.62% :no_entry:
Sema.AccessLevelRequest | 18,775,402 | 19,128,420 | 353,017 | 1.88% :no_entry:
Sema.AttachedPropertyWrappersRequest | 22,945,745 | 23,198,778 | 253,033 | 1.1% :no_entry:
Sema.CollectOverriddenDeclsRequest | 15,173,002 | 15,469,387 | 296,385 | 1.95% :no_entry:
Sema.LookupAllConformancesInContextRequest | 3,148,295 | 3,189,023 | 40,728 | 1.29% :no_entry:
Sema.PrimaryAssociatedTypesRequest | 19,493 | 19,695 | 202 | 1.04% :no_entry:
Sema.ProvideDefaultImplForRequest | 15,173,002 | 15,469,387 | 296,385 | 1.95% :no_entry:
Sema.SimpleDidSetRequest | 3,723,513 | 3,791,746 | 68,233 | 1.83% :no_entry:
Sema.USRGenerationRequest | 16,910,985 | 17,213,874 | 302,889 | 1.79% :no_entry:
</details>
<details>
<summary>Improved (3)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.ModuleShadowCacheHit | 56,971 | 55,510 | -1,461 | -2.56% :white_check_mark:
AST.ModuleShadowCacheMiss | 4,494 | 4,426 | -68 | -1.51% :white_check_mark:
Frontend.NumInstructionsExecuted | 145,026,849,632,443 | 122,931,368,255,179 | -22,095,481,377,264 | -15.24% :white_check_mark:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (308)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.ImportSetCacheHit | 4,968,569 | 4,968,810 | 241 | 0.0%
AST.ImportSetCacheMiss | 793,969 | 793,971 | 1 | 0.0%
AST.ImportSetFoldHit | 402,528 | 402,505 | -23 | -0.01%
AST.ImportSetFoldMiss | 391,441 | 391,466 | 25 | 0.01%
AST.ModuleVisibilityCacheHit | 27,662,049 | 27,662,049 | 0 | 0.0%
AST.ModuleVisibilityCacheMiss | 209,685 | 209,685 | 0 | 0.0%
AST.NumASTBytesAllocated | 66,050,072,262 | 66,242,215,133 | 192,142,870 | 0.29%
AST.NumASTScopeExpansions | 8,308,002 | 8,308,205 | 203 | 0.0%
AST.NumASTScopeLookups | 9,532,495 | 9,533,065 | 569 | 0.01%
AST.NumDecls | 340,649 | 340,649 | 0 | 0.0%
AST.NumDependencies | 441,698 | 441,699 | 1 | 0.0%
AST.NumIncrementalDependencies | 25,214 | 25,214 | 0 | 0.0%
AST.NumLinkLibraries | 0 | 0 | 0 | 0.0%
AST.NumLoadedModules | 322,695 | 322,695 | 0 | 0.0%
AST.NumLocalTypeDecls | 6,522 | 6,522 | 0 | 0.0%
AST.NumModuleLookupClassMember | 6,798 | 6,798 | 0 | 0.0%
AST.NumModuleLookupValue | 163,313,265 | 163,311,373 | -1,891 | -0.0%
AST.NumObjCMethods | 93,972 | 93,972 | 0 | 0.0%
AST.NumOperators | 1,028 | 1,028 | 0 | 0.0%
AST.NumPrecedenceGroups | 176 | 176 | 0 | 0.0%
AST.NumReferencedDynamicNames | 159 | 159 | 0 | 0.0%
AST.NumReferencedMemberNames | 14,969,282 | 14,969,274 | -8 | -0.0%
AST.NumReferencedTopLevelNames | 1,342,507 | 1,342,507 | 0 | 0.0%
AST.NumSourceBuffers | 395,204 | 395,204 | 0 | 0.0%
AST.NumSourceLines | 12,356,688 | 12,356,688 | 0 | 0.0%
AST.NumTotalClangImportedEntities | 1,954,415 | 1,956,895 | 2,480 | 0.13%
Frontend.MaxMallocUsage | 0 | 0 | 0 | 0.0%
Frontend.NumProcessFailures | 31 | 31 | 0 | 0.0%
IRGen.IRGenRequest | 29,478 | 29,478 | 0 | 0.0%
IRGen.OptimizedIRRequest | 0 | 0 | 0 | 0.0%
IRGen.SymbolObjectCodeRequest | 0 | 0 | 0 | 0.0%
IRModule.NumGOTEntries | 297,857 | 297,857 | 0 | 0.0%
IRModule.NumIRAliases | 239,487 | 239,487 | 0 | 0.0%
IRModule.NumIRBasicBlocks | 8,261,723 | 8,261,723 | 0 | 0.0%
IRModule.NumIRComdatSymbols | 0 | 0 | 0 | 0.0%
IRModule.NumIRFunctions | 4,497,009 | 4,497,009 | 0 | 0.0%
IRModule.NumIRGlobals | 4,273,592 | 4,273,592 | 0 | 0.0%
IRModule.NumIRIFuncs | 0 | 0 | 0 | 0.0%
IRModule.NumIRInsts | 105,576,298 | 105,576,298 | 0 | 0.0%
IRModule.NumIRNamedMetaData | 119,432 | 119,432 | 0 | 0.0%
IRModule.NumIRValueSymbols | 8,842,460 | 8,842,460 | 0 | 0.0%
LLVM.NumLLVMBytesOutput | 2,443,809,896 | 2,443,852,960 | 43,064 | 0.0%
Parse.IDEInspectionSecondPassRequest | 0 | 0 | 0 | 0.0%
Parse.NumFunctionsParsed | 396,751 | 396,751 | 0 | 0.0%
Parse.NumIterableDeclContextParsed | 805,328 | 805,334 | 6 | 0.0%
Parse.ParseAbstractFunctionBodyRequest | 684,546 | 684,546 | 0 | 0.0%
Parse.ParseMembersRequest | 617,180 | 617,186 | 6 | 0.0%
Parse.ParseSourceFileRequest | 369,509 | 369,509 | 0 | 0.0%
Parse.ParseTopLevelDeclsRequest | 369,509 | 369,509 | 0 | 0.0%
SILGen.ASTLoweringRequest | 30,813 | 30,813 | 0 | 0.0%
SILGen.ParseSILModuleRequest | 0 | 0 | 0 | 0.0%
SILModule.NumSILGenDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILGenFunctions | 3,511,580 | 3,511,580 | 0 | 0.0%
SILModule.NumSILGenGlobalVariables | 140,013 | 140,013 | 0 | 0.0%
SILModule.NumSILGenVtables | 40,379 | 40,379 | 0 | 0.0%
SILModule.NumSILGenWitnessTables | 197,169 | 197,169 | 0 | 0.0%
SILModule.NumSILOptDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILOptFunctions | 4,011,125 | 4,011,125 | 0 | 0.0%
SILModule.NumSILOptGlobalVariables | 142,008 | 142,008 | 0 | 0.0%
SILModule.NumSILOptVtables | 40,742 | 40,742 | 0 | 0.0%
SILModule.NumSILOptWitnessTables | 240,557 | 240,557 | 0 | 0.0%
SILOptimizer.ExecuteSILPipelineRequest | 151,423 | 151,423 | 0 | 0.0%
SILOptimizer.LoweredSILRequest | 0 | 0 | 0 | 0.0%
Sema.ABIMembersRequest | 309,274 | 309,274 | 0 | 0.0%
Sema.AbstractGenericSignatureRequest | 145,559 | 145,970 | 410 | 0.28%
Sema.ActorIsolationRequest | 7,118,667 | 7,119,546 | 879 | 0.01%
Sema.AllMembersRequest | 117,897 | 117,897 | 0 | 0.0%
Sema.AnyObjectLookupRequest | 234 | 234 | 0 | 0.0%
Sema.ApplyAccessNoteRequest | 4,363,854 | 4,366,559 | 2,705 | 0.06%
Sema.AreAllStoredPropertiesDefaultInitableRequest | 49,574 | 49,574 | 0 | 0.0%
Sema.AttachedPropertyWrapperTypeRequest | 30,646 | 30,646 | 0 | 0.0%
Sema.AttachedResultBuilderRequest | 1,880,502 | 1,880,502 | 0 | 0.0%
Sema.BodyInitKindRequest | 94,303 | 94,303 | 0 | 0.0%
Sema.BreakTargetRequest | 6,567 | 6,567 | 0 | 0.0%
Sema.CXXNamespaceMemberLookup | 0 | 0 | 0 | 0.0%
Sema.CallerSideDefaultArgExprRequest | 479,669 | 479,669 | 0 | 0.0%
Sema.CheckDistributedFunctionRequest | 80 | 80 | 0 | 0.0%
Sema.CheckInconsistentAccessLevelOnImport | 59,478 | 59,478 | 0 | 0.0%
Sema.CheckInconsistentImplementationOnlyImportsRequest | 10,759 | 10,759 | 0 | 0.0%
Sema.CheckInconsistentSPIOnlyImportsRequest | 59,478 | 59,478 | 0 | 0.0%
Sema.CheckInconsistentWeakLinkedImportsRequest | 10,759 | 10,759 | 0 | 0.0%
Sema.CheckRedeclarationRequest | 2,331,978 | 2,331,978 | 0 | 0.0%
Sema.ClangCategoryLookupRequest | 0 | 0 | 0 | 0.0%
Sema.ClangDirectLookupRequest | 15,213 | 15,213 | 0 | 0.0%
Sema.ClangRecordMemberLookup | 15,213 | 15,213 | 0 | 0.0%
Sema.ClassAncestryFlagsRequest | 152,942 | 153,162 | 219 | 0.14%
Sema.ClosureEffectsRequest | 152,909 | 152,909 | 0 | 0.0%
Sema.ClosureHasExplicitResultRequest | 58,742 | 58,742 | 0 | 0.0%
Sema.CompareDeclSpecializationRequest | 870,623 | 871,970 | 1,347 | 0.15%
Sema.CompilerPluginLoadRequest | 0 | 0 | 0 | 0.0%
Sema.ConditionalRequirementsRequest | 876,356 | 879,184 | 2,827 | 0.32%
Sema.ConformanceHasEffectRequest | 63 | 63 | 0 | 0.0%
Sema.ConstantValueInfoRequest | 0 | 0 | 0 | 0.0%
Sema.ContinueTargetRequest | 2,180 | 2,180 | 0 | 0.0%
Sema.CursorInfoRequest | 0 | 0 | 0 | 0.0%
Sema.CustomAttrNominalRequest | 22,083 | 22,083 | 0 | 0.0%
Sema.CustomAttrTypeRequest | 3,280 | 3,280 | 0 | 0.0%
Sema.CustomRefCountingOperation | 0 | 0 | 0 | 0.0%
Sema.CxxRecordAsSwiftType | 0 | 0 | 0 | 0.0%
Sema.CxxRecordSemantics | 85,514 | 85,605 | 91 | 0.11%
Sema.DefaultAndMaxAccessLevelRequest | 121,658 | 121,658 | 0 | 0.0%
Sema.DefaultArgumentExprRequest | 103,047 | 103,047 | 0 | 0.0%
Sema.DefaultArgumentInitContextRequest | 841 | 841 | 0 | 0.0%
Sema.DefaultArgumentTypeRequest | 21,688 | 21,688 | 0 | 0.0%
Sema.DefaultDefinitionTypeRequest | 8,506 | 8,506 | 0 | 0.0%
Sema.DefaultTypeRequest | 655,408 | 655,408 | 0 | 0.0%
Sema.DerivativeAttrOriginalDeclRequest | 0 | 0 | 0 | 0.0%
Sema.DifferentiableAttributeTypeCheckRequest | 0 | 0 | 0 | 0.0%
Sema.DirectLookupRequest | 80,958,563 | 81,087,365 | 128,802 | 0.16%
Sema.DirectOperatorLookupRequest | 2,577,648 | 2,576,481 | -1,167 | -0.05%
Sema.DirectPrecedenceGroupLookupRequest | 951,307 | 951,307 | 0 | 0.0%
Sema.DistributedModuleIsAvailableRequest | 67 | 67 | 0 | 0.0%
Sema.DynamicallyReplacedDeclRequest | 7,566,094 | 7,566,970 | 876 | 0.01%
Sema.EnumElementExprPatternRequest | 17 | 17 | 0 | 0.0%
Sema.EnumRawTypeRequest | 41,362 | 41,362 | 0 | 0.0%
Sema.EnumRawValuesRequest | 24,332 | 24,332 | 0 | 0.0%
Sema.ExistentialConformsToSelfRequest | 33,909 | 33,890 | -19 | -0.06%
Sema.ExpandAccessorMacros | 1,303,150 | 1,303,150 | 0 | 0.0%
Sema.ExpandConformanceMacros | 1,292,664 | 1,297,000 | 4,335 | 0.34%
Sema.ExpandMacroExpansionDeclRequest | 0 | 0 | 0 | 0.0%
Sema.ExpandMacroExpansionExprRequest | 0 | 0 | 0 | 0.0%
Sema.ExpandMemberAttributeMacros | 74,722,659 | 75,413,625 | 690,966 | 0.92%
Sema.ExpandPeerMacroRequest | 10,278,153 | 10,278,153 | 0 | 0.0%
Sema.ExpandSynthesizedMemberMacroRequest | 303,199 | 303,199 | 0 | 0.0%
Sema.ExprPatternMatchRequest | 92,535 | 92,535 | 0 | 0.0%
Sema.ExtendedNominalRequest | 706,326 | 706,326 | 0 | 0.0%
Sema.ExtendedTypeRequest | 217,307 | 217,307 | 0 | 0.0%
Sema.ExternalMacroDefinitionRequest | 0 | 0 | 0 | 0.0%
Sema.FragileFunctionKindRequest | 2,791,267 | 2,791,267 | 0 | 0.0%
Sema.FunctionOperatorRequest | 20,338 | 20,338 | 0 | 0.0%
Sema.GenericParamListRequest | 7,387,600 | 7,400,295 | 12,695 | 0.17%
Sema.GenericSignatureRequest | 3,622,991 | 3,626,150 | 3,159 | 0.09%
Sema.GetDestructorRequest | 47,188 | 47,188 | 0 | 0.0%
Sema.GetDistributedActorArgumentDecodingMethodRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedActorIDPropertyRequest | 72 | 72 | 0 | 0.0%
Sema.GetDistributedActorImplicitCodableRequest | 117 | 117 | 0 | 0.0%
Sema.GetDistributedActorInvocationDecoderRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedActorSystemPropertyRequest | 63 | 63 | 0 | 0.0%
Sema.GetDistributedActorSystemRemoteCallFunctionRequest | 46 | 46 | 0 | 0.0%
Sema.GetDistributedRemoteCallArgumentInitFunctionRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedRemoteCallTargetInitFunctionRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedTargetInvocationDecoderDecodeNextArgumentFunctionRequest | 22 | 22 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordArgumentFunctionRequest | 20 | 20 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordErrorTypeFunctionRequest | 11 | 11 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordReturnTypeFunctionRequest | 16 | 16 | 0 | 0.0%
Sema.GetDistributedTargetInvocationResultHandlerOnReturnFunctionRequest | 12 | 12 | 0 | 0.0%
Sema.GetDistributedThunkRequest | 80 | 80 | 0 | 0.0%
Sema.GetImplicitSendableRequest | 265,055 | 263,360 | -1,695 | -0.64%
Sema.GetRuntimeDiscoverableAttributes | 2,453,177 | 2,453,177 | 0 | 0.0%
Sema.GetSourceFileAsyncNode | 138 | 138 | 0 | 0.0%
Sema.GlobalActorAttributeRequest | 9,041,559 | 9,043,185 | 1,626 | 0.02%
Sema.GlobalActorInstanceRequest | 1,063 | 1,063 | 0 | 0.0%
Sema.HasCircularInheritedProtocolsRequest | 18,513 | 18,513 | 0 | 0.0%
Sema.HasCircularRawValueRequest | 25,096 | 25,096 | 0 | 0.0%
Sema.HasDefaultInitRequest | 106,862 | 106,862 | 0 | 0.0%
Sema.HasDynamicCallableAttributeRequest | 31,859 | 31,859 | 0 | 0.0%
Sema.HasDynamicMemberLookupAttributeRequest | 581,014 | 581,739 | 725 | 0.12%
Sema.HasImportsMatchingFlagRequest | 733,008 | 733,008 | 0 | 0.0%
Sema.HasIsolatedSelfRequest | 15,938,580 | 16,041,413 | 102,833 | 0.65%
Sema.HasMemberwiseInitRequest | 52,221 | 52,221 | 0 | 0.0%
Sema.HasMissingDesignatedInitializersRequest | 33,330 | 33,330 | 0 | 0.0%
Sema.HasSelfOrAssociatedTypeRequirementsRequest | 16,730 | 16,730 | 0 | 0.0%
Sema.HasUserDefinedDesignatedInitRequest | 106,872 | 106,872 | 0 | 0.0%
Sema.IDEInspectionFileRequest | 0 | 0 | 0 | 0.0%
Sema.InferredGenericSignatureRequest | 242,028 | 242,028 | 0 | 0.0%
Sema.InheritedDeclsReferencedRequest | 7,816,620 | 7,877,659 | 61,039 | 0.78%
Sema.InheritedProtocolsRequest | 865,895 | 866,569 | 674 | 0.08%
Sema.InheritedTypeRequest | 473,445 | 473,461 | 16 | 0.0%
Sema.InheritsSuperclassInitializersRequest | 42,292 | 42,292 | 0 | 0.0%
Sema.InitKindRequest | 197,260 | 197,260 | 0 | 0.0%
Sema.InterfaceTypeRequest | 24,458,480 | 24,702,333 | 243,852 | 1.0%
Sema.IsABICompatibleOverrideRequest | 322,051 | 322,051 | 0 | 0.0%
Sema.IsAccessorTransparentRequest | 576,041 | 576,041 | 0 | 0.0%
Sema.IsActorRequest | 2,521,455 | 2,525,095 | 3,640 | 0.14%
Sema.IsCallAsFunctionNominalRequest | 6,761 | 6,761 | 0 | 0.0%
Sema.IsDeclApplicableRequest | 0 | 0 | 0 | 0.0%
Sema.IsDeclRefinementOfRequest | 71,755 | 71,755 | 0 | 0.0%
Sema.IsDefaultActorRequest | 63,944 | 63,944 | 0 | 0.0%
Sema.IsDistributedActorRequest | 2,807,705 | 2,812,274 | 4,569 | 0.16%
Sema.IsDynamicRequest | 3,285,947 | 3,285,947 | 0 | 0.0%
Sema.IsFinalRequest | 5,003,230 | 5,038,396 | 35,166 | 0.7%
Sema.IsGetterMutatingRequest | 827,800 | 827,800 | 0 | 0.0%
Sema.IsImplicitlyUnwrappedOptionalRequest | 4,076,002 | 4,075,674 | -327 | -0.01%
Sema.IsMoveOnlyRequest | 781,894 | 782,244 | 349 | 0.04%
Sema.IsNonUserModuleRequest | 104,502 | 104,355 | -147 | -0.14%
Sema.IsObjCRequest | 3,056,827 | 3,059,532 | 2,705 | 0.09%
Sema.IsSafeUseOfCxxDecl | 0 | 0 | 0 | 0.0%
Sema.IsSetterMutatingRequest | 742,425 | 742,425 | 0 | 0.0%
Sema.IsSingleValueStmtRequest | 374,454 | 374,454 | 0 | 0.0%
Sema.IsStaticRequest | 1,660,833 | 1,661,096 | 262 | 0.02%
Sema.LazyStoragePropertyRequest | 2,573 | 2,573 | 0 | 0.0%
Sema.LocalDiscriminatorsRequest | 551,530 | 551,530 | 0 | 0.0%
Sema.LookupConformanceInModuleRequest | 128,500,220 | 128,983,031 | 482,811 | 0.38%
Sema.LookupInModuleRequest | 11,145,919 | 11,141,887 | -4,032 | -0.04%
Sema.LookupInfixOperatorRequest | 82,322 | 82,322 | 0 | 0.0%
Sema.LookupPostfixOperatorRequest | 74 | 74 | 0 | 0.0%
Sema.LookupPrecedenceGroupRequest | 33,550 | 33,550 | 0 | 0.0%
Sema.LookupPrefixOperatorRequest | 452 | 452 | 0 | 0.0%
Sema.MacroDefinitionRequest | 0 | 0 | 0 | 0.0%
Sema.MangleLocalTypeDeclRequest | 3,149 | 3,149 | 0 | 0.0%
Sema.ModuleImplicitImportsRequest | 10,759 | 10,759 | 0 | 0.0%
Sema.ModuleLibraryLevelRequest | 22 | 22 | 0 | 0.0%
Sema.ModuleQualifiedLookupRequest | 4,055,233 | 4,050,632 | -4,601 | -0.11%
Sema.NamedLazyMemberLoadSuccessCount | 37,940,480 | 37,923,536 | -16,944 | -0.04%
Sema.NamingPatternRequest | 242,249 | 242,249 | 0 | 0.0%
Sema.NeedsNewVTableEntryRequest | 1,072,248 | 1,072,248 | 0 | 0.0%
Sema.NumAccessorBodiesSynthesized | 480,199 | 480,199 | 0 | 0.0%
Sema.NumAccessorsSynthesized | 579,269 | 579,269 | 0 | 0.0%
Sema.NumConformancePathsRecorded | 996,792 | 997,326 | 534 | 0.05%
Sema.NumConformancesDeserialized | 6,238,530 | 6,290,075 | 51,544 | 0.83%
Sema.NumConstraintScopes | 80,307,034 | 80,321,129 | 14,094 | 0.02%
Sema.NumConstraintsConsideredForEdgeContraction | 2,726,255 | 2,726,255 | 0 | 0.0%
Sema.NumCrossImportsChecked | 0 | 0 | 0 | 0.0%
Sema.NumCrossImportsFound | 0 | 0 | 0 | 0.0%
Sema.NumCyclicOneWayComponentsCollapsed | 0 | 0 | 0 | 0.0%
Sema.NumDeclsDeserialized | 119,615,768 | 120,342,035 | 726,266 | 0.61%
Sema.NumDeclsTypechecked | 3,314,288 | 3,314,288 | 0 | 0.0%
Sema.NumLazyIterableDeclContexts | 15,587,344 | 15,603,145 | 15,801 | 0.1%
Sema.NumLazyRequirementSignatures | 1,191,573 | 1,191,956 | 383 | 0.03%
Sema.NumLazyRequirementSignaturesLoaded | 792,157 | 792,750 | 592 | 0.07%
Sema.NumLeafScopes | 61,816,346 | 61,828,324 | 11,977 | 0.02%
Sema.NumRequirementMachineCompletionSteps | 29,848,554 | 30,011,691 | 163,137 | 0.55%
Sema.NumRequirementMachineUnifiedConcreteTerms | 1,369,732 | 1,372,571 | 2,838 | 0.21%
Sema.NumRequirementMachines | 2,341,034 | 2,349,659 | 8,625 | 0.37%
Sema.NumTypesDeserialized | 32,961,860 | 33,039,154 | 77,293 | 0.23%
Sema.NumUnloadedLazyIterableDeclContexts | 8,809,499 | 8,786,115 | -23,383 | -0.27%
Sema.ObjCInterfaceAndImplementationRequest | 949,355 | 949,352 | -2 | -0.0%
Sema.OpaqueReadOwnershipRequest | 502,884 | 502,884 | 0 | 0.0%
Sema.OpaqueResultTypeRequest | 513 | 513 | 0 | 0.0%
Sema.OperatorPrecedenceGroupRequest | 1,027 | 1,027 | 0 | 0.0%
Sema.OverriddenDeclsRequest | 6,659,772 | 6,674,377 | 14,605 | 0.22%
Sema.ParamSpecifierRequest | 2,243,953 | 2,243,953 | 0 | 0.0%
Sema.PatternBindingEntryRequest | 817,936 | 817,936 | 0 | 0.0%
Sema.PatternTypeRequest | 969,967 | 969,967 | 0 | 0.0%
Sema.PolymorphicEffectKindRequest | 1,121,023 | 1,121,023 | 0 | 0.0%
Sema.PolymorphicEffectRequirementsRequest | 75 | 75 | 0 | 0.0%
Sema.PreCheckResultBuilderRequest | 951 | 951 | 0 | 0.0%
Sema.PreCheckReturnStmtRequest | 595,368 | 595,368 | 0 | 0.0%
Sema.PrimarySourceFilesRequest | 10,759 | 10,759 | 0 | 0.0%
Sema.PropertyWrapperAuxiliaryVariablesRequest | 2,452,273 | 2,452,273 | 0 | 0.0%
Sema.PropertyWrapperBackingPropertyTypeRequest | 30,646 | 30,646 | 0 | 0.0%
Sema.PropertyWrapperInitializerInfoRequest | 715,935 | 715,935 | 0 | 0.0%
Sema.PropertyWrapperLValuenessRequest | 2,636 | 2,636 | 0 | 0.0%
Sema.PropertyWrapperMutabilityRequest | 821,669 | 821,669 | 0 | 0.0%
Sema.PropertyWrapperTypeInfoRequest | 599 | 599 | 0 | 0.0%
Sema.ProtocolDependenciesRequest | 678,535 | 679,022 | 487 | 0.07%
Sema.ProtocolRequirementsRequest | 81,344 | 81,344 | 0 | 0.0%
Sema.ProtocolRequiresClassRequest | 47,887 | 47,920 | 32 | 0.07%
Sema.QualifiedLookupRequest | 10,031,616 | 10,050,674 | 19,058 | 0.19%
Sema.RangeInfoRequest | 0 | 0 | 0 | 0.0%
Sema.RawCommentRequest | 1,185,057 | 1,185,057 | 0 | 0.0%
Sema.RenamedDeclRequest | 5,221 | 5,221 | 0 | 0.0%
Sema.RequirementRequest | 194,117 | 194,117 | 0 | 0.0%
Sema.RequirementSignatureRequest | 851,585 | 852,229 | 644 | 0.08%
Sema.RequiresOpaqueAccessorsRequest | 2,100,358 | 2,100,358 | 0 | 0.0%
Sema.RequiresOpaqueModifyCoroutineRequest | 508,305 | 508,305 | 0 | 0.0%
Sema.ResolveEffectiveMemberwiseInitRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveImplicitMemberRequest | 947,225 | 949,992 | 2,766 | 0.29%
Sema.ResolveMacroRequest | 24,989 | 24,989 | 0 | 0.0%
Sema.ResolveProtocolNameRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveTypeEraserTypeRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveTypeRequest | 5,759,029 | 5,759,029 | 0 | 0.0%
Sema.ResultBuilderTypeRequest | 555,155 | 555,155 | 0 | 0.0%
Sema.ResultTypeRequest | 1,259,221 | 1,259,221 | 0 | 0.0%
Sema.RootAndResultTypeOfKeypathDynamicMemberRequest | 0 | 0 | 0 | 0.0%
Sema.RootTypeOfKeypathDynamicMemberRequest | 0 | 0 | 0 | 0.0%
Sema.SPIGroupsRequest | 8,790,286 | 8,818,105 | 27,819 | 0.32%
…